# api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Manager Service
    description: This service manages the ADs
    version: 0.0.1
servers:
    - url: http://localhost:8000
paths:
    /ad:
        get:
            tags:
                - manager
            summary: List all ADs by filter
            description: List all ADs by filter
            operationId: manager#list_ads
            parameters:
                - name: offset
                  in: query
                  description: Offset of AD
                  allowEmptyValue: true
                  required: true
                  schema:
                    type: integer
                    description: Offset of AD
                    default: 0
                    example: 0
                    format: int64
                    minimum: 0
                  example: 0
                - name: limit
                  in: query
                  description: Limit of AD
                  allowEmptyValue: true
                  required: true
                  schema:
                    type: integer
                    description: Limit of AD
                    default: 5
                    example: 10
                    format: int64
                    minimum: 1
                  example: 10
                - name: age_start
                  in: query
                  description: Start age of target
                  allowEmptyValue: true
                  schema:
                    type: integer
                    description: Start age of target
                    example: 18
                    format: int64
                    minimum: 1
                    maximum: 100
                  example: 18
                - name: age_end
                  in: query
                  description: End age of target
                  allowEmptyValue: true
                  schema:
                    type: integer
                    description: End age of target
                    example: 60
                    format: int64
                    minimum: 1
                    maximum: 100
                  example: 60
                - name: gender
                  in: query
                  description: Gender of target
                  allowEmptyValue: true
                  schema:
                    type: array
                    items:
                        type: string
                        example: F
                        pattern: M|F
                    description: Gender of target
                    example:
                        - M
                        - F
                  example:
                    - M
                    - F
                - name: country
                  in: query
                  description: Country of target
                  allowEmptyValue: true
                  schema:
                    type: array
                    items:
                        type: string
                        example: JP
                        pattern: TW|JP
                    description: Country of target
                    example:
                        - TW
                        - JP
                  example:
                    - TW
                    - JP
                - name: platform
                  in: query
                  description: Platform of target
                  allowEmptyValue: true
                  schema:
                    type: array
                    items:
                        type: string
                        example: ios
                        pattern: ios|android|web
                    description: Platform of target
                    example:
                        - ios
                        - android
                        - web
                  example:
                    - ios
                    - android
                    - web
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Ad'
                                example:
                                    - end_at: "2024-12-10T03:00:00.000Z"
                                      title: AD 1
                                    - end_at: "2024-12-10T03:00:00.000Z"
                                      title: AD 1
                            example:
                                - end_at: "2024-12-10T03:00:00.000Z"
                                  title: AD 1
                                - end_at: "2024-12-10T03:00:00.000Z"
                                  title: AD 1
                                - end_at: "2024-12-10T03:00:00.000Z"
                                  title: AD 1
        post:
            tags:
                - manager
            summary: Create a new AD
            description: Create a new edge
            operationId: manager#create_ad
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAdRequestBody'
                        example:
                            conditions:
                                age_end: 60
                                age_start: 18
                                country:
                                    - TW
                                    - JP
                                gender:
                                    - M
                                    - F
                                platform:
                                    - ios
                                    - android
                                    - web
                            end_at: "2024-12-10T03:00:00.000Z"
                            start_at: "2024-03-10T03:00:00.000Z"
                            title: AD 1
            responses:
                "201":
                    description: Created response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateAdResponseBody'
                            example:
                                id: "34"
components:
    schemas:
        Ad:
            type: object
            properties:
                end_at:
                    type: string
                    description: End time of AD
                    example: "2024-12-10T03:00:00.000Z"
                    format: date-time
                title:
                    type: string
                    description: Title of AD
                    example: AD 1
                    minLength: 1
                    maxLength: 100
            description: List all ads by filter
            example:
                end_at: "2024-12-10T03:00:00.000Z"
                title: AD 1
            required:
                - title
                - end_at
        CreateAdRequestBody:
            type: object
            properties:
                conditions:
                    type: object
                    properties:
                        age_end:
                            type: integer
                            description: End age of target
                            example: 60
                            format: int64
                            minimum: 1
                            maximum: 100
                        age_start:
                            type: integer
                            description: Start age of target
                            example: 18
                            format: int64
                            minimum: 1
                            maximum: 100
                        country:
                            type: array
                            items:
                                type: string
                                example: JP
                                pattern: TW|JP
                            description: Country of target
                            example:
                                - TW
                                - JP
                        gender:
                            type: array
                            items:
                                type: string
                                example: M
                                pattern: M|F
                            description: Gender of target
                            example:
                                - M
                                - F
                        platform:
                            type: array
                            items:
                                type: string
                                example: ios
                                pattern: ios|android|web
                            description: Platform of target
                            example:
                                - ios
                                - android
                                - web
                    example:
                        age_end: 60
                        age_start: 18
                        country:
                            - TW
                            - JP
                        gender:
                            - M
                            - F
                        platform:
                            - ios
                            - android
                            - web
                end_at:
                    type: string
                    description: End time of AD
                    example: "2024-12-10T03:00:00.000Z"
                    format: date-time
                start_at:
                    type: string
                    description: Start time of AD
                    example: "2024-03-10T03:00:00.000Z"
                    format: date-time
                title:
                    type: string
                    description: Title of AD
                    example: AD 1
                    minLength: 1
                    maxLength: 100
            example:
                conditions:
                    age_end: 60
                    age_start: 18
                    country:
                        - TW
                        - JP
                    gender:
                        - M
                        - F
                    platform:
                        - ios
                        - android
                        - web
                end_at: "2024-12-10T03:00:00.000Z"
                start_at: "2024-03-10T03:00:00.000Z"
                title: AD 1
            required:
                - title
                - start_at
                - end_at
        CreateAdResponseBody:
            type: object
            properties:
                id:
                    type: string
                    description: ID of the AD
                    example: "34"
            example:
                id: "34"
            required:
                - id
tags:
    - name: manager